Initial commit of ICBR
This is a re-do on Minecraft 1.10.2. I was unable to get textures to
load at all on 1.7.10, so I started over. This includes a few basic
things-- I went ahead and did a couple of basic items and a basic block
in order to test the viability of doing things the new way on 1.10.2.

The basic items/blocks are using textures from Shadowfacts' tutorials,
with the exception of the copperCoil, which is currently using a
texture from IC2 (it will be changed to something original soon). The
current textures are simply placeholders while I try to figure out how
to do the basics.