Add a texture for the ring launcher
Made a texture and figured out how to put it on the custom model. Still
using the stone brick texture for the top and bottom of the wider parts
of the launcher, but I haven't yet figured out how to get the texture
space for that large of a surface area.