Added new Ring item and ability to put it on the launcher
Created a new model for the Ring and textures to go with it. Made the
model with Blockbench (blockbench.net) (really nice, by the way; made
figuring out how to make a ring extremely easy, after I spent an hour or
two trying to puzzle out how to do it by hand).

Also made the RingLauncher a TileEntity and added the corresponding code
to allow items to be placed on it. Made it so the item that was
placeable was the new Ring, and made it so a player could put a single
Ring onto it and take it off again, with some handling for adding it to
the stack of Rings the player may be holding.

Credit to Shadowfacts (https://shadowfacts.net/tutorials) for the
tutorials on how to get this far, they were very helpful.